Move list item components into a new folder
