+++ .eslintignore
@@ -395,16 +395,8 @@ packages/app-desktop/gui/Root_UpgradeSyncTarget.js
 packages/app-desktop/gui/SearchBar/SearchBar.js
 packages/app-desktop/gui/ShareFolderDialog/ShareFolderDialog.js
 packages/app-desktop/gui/ShareNoteDialog.js
-packages/app-desktop/gui/Sidebar/AllNotesItem.js
-packages/app-desktop/gui/Sidebar/EmptyExpandLink.js
-packages/app-desktop/gui/Sidebar/ExpandIcon.js
-packages/app-desktop/gui/Sidebar/ExpandLink.js
-packages/app-desktop/gui/Sidebar/FolderItem.js
-packages/app-desktop/gui/Sidebar/HeaderItem.js
-packages/app-desktop/gui/Sidebar/NoteCount.js
 packages/app-desktop/gui/Sidebar/NotebookAndTagList.js
 packages/app-desktop/gui/Sidebar/Sidebar.js
-packages/app-desktop/gui/Sidebar/TagItem.js
 packages/app-desktop/gui/Sidebar/commands/focusElementSideBar.js
 packages/app-desktop/gui/Sidebar/commands/index.js
 packages/app-desktop/gui/Sidebar/hooks/useFocusHandler.js
@@ -412,6 +404,14 @@ packages/app-desktop/gui/Sidebar/hooks/useOnRenderItem.js
 packages/app-desktop/gui/Sidebar/hooks/useOnSidebarKeyDownHandler.js
 packages/app-desktop/gui/Sidebar/hooks/useSelectedSidebarIndex.js
 packages/app-desktop/gui/Sidebar/hooks/useSidebarListData.js
+packages/app-desktop/gui/Sidebar/listItemComponents/AllNotesItem.js
+packages/app-desktop/gui/Sidebar/listItemComponents/EmptyExpandLink.js
+packages/app-desktop/gui/Sidebar/listItemComponents/ExpandIcon.js
+packages/app-desktop/gui/Sidebar/listItemComponents/ExpandLink.js
+packages/app-desktop/gui/Sidebar/listItemComponents/FolderItem.js
+packages/app-desktop/gui/Sidebar/listItemComponents/HeaderItem.js
+packages/app-desktop/gui/Sidebar/listItemComponents/NoteCount.js
+packages/app-desktop/gui/Sidebar/listItemComponents/TagItem.js
 packages/app-desktop/gui/Sidebar/styles/index.js
 packages/app-desktop/gui/Sidebar/types.js
 packages/app-desktop/gui/StatusScreen/StatusScreen.js
