+++ packages/app-desktop/gui/Sidebar/listItemComponents/ExpandIcon.tsx
