+++ packages/app-desktop/gui/Sidebar/listItemComponents/NoteCount.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { StyledNoteCount } from './styles';
+import { StyledNoteCount } from '../styles';
 
 
 interface Props {
